+++ totalRP3/Modules/Dashboard/WhatsNewTabView.lua
@@ -22,12 +22,7 @@ local strhtml = TRP3_API.utils.str.toHTML;
 --- List of "What's new" segments to concatenate and display together when
 ---  this view is activated.
 local SEGMENTS = {
-	loc.WHATS_NEW_27_5,
-	loc.WHATS_NEW_27_4,
-	loc.WHATS_NEW_27_3,
-	loc.WHATS_NEW_27_2,
-	loc.WHATS_NEW_27_1,
-	loc.WHATS_NEW_27_0,
+	loc.WHATS_NEW_28_0,
 };
 
 --- Returns the fully formatted localized text for this view.